refactor(RacingVariables.js): make class to file
