+++ src/constants/RacingVariables.js
@@ -1,0 +1,10 @@
+const RACING_VARIABLES = {
+    INPUT_PROMPT : "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)\n",
+    INPUT_COUNT_PROMPT : "시도할 횟수는 몇 회인가요?\n",
+    POSITION_MARKER : "-",
+    MAXIMUM_ROUND : 10,
+    WINNER_PROMPT : "최종 우승자"
+  
+  }
+
+  export default RACING_VARIABLES;
